Implemented text-based number extraction
A new method 'extract_number' was added to the 'SceneDescriptionGenerator' class which extracts the first number from a given text string. This is used in the 'retrieve_scene_number' method, which now repeatedly retries a conversation process (up to a maximum number of times) to successfully retrieve a valid number from a response.