Update SceneDescriptionGenerator for comic panels
The SceneDescriptionGenerator class in the echo-artistry project has been refactored to work with comic panels instead of movie scenes. The concept of "scenes" has been replaced with "panels" to better suit a comic context. The CompositeOption enum class has been added to manage the number of images in each comic. The methods in the SceneDescriptionGenerator class have been updated to work with these changes, including changing the method names and adjusting the text processing.