Refactor application to generate comic stories
Introduced new functionality to the echo_artistry application which generates comic stories based on an inputted voice memo. Renamed SceneDescriptionGenerator class to ComicStoryGenerator. The Transcriber class was updated to handle audio file transcription in a more streamlined manner. A new feature was also added to allow the user to specify an audio file path as an argument.