Refactor CostManager and OpenAIClient classes
This commit moves the instantiation of the CostManager class to the constructor of OpenAIClient. Also, 'total_cost' is now an attribute of the CostManager class, not an instance variable. This refactoring tightens up the code and makes the cost calculation process more straightforward, as all the needed components are now present at the time of OpenAIClient's creation.