Add GitHub Actions workflow for package publication (#12)
* Add GitHub Actions workflow for package publication

Added a new GitHub Actions workflow (release.yaml) for automatic publishing of the package to PyPI upon publishing a release. In addition, updates are made to the pyproject.toml for package publishing requirements and to the README.md for correct function calling instructions.

* Update documentation link in README

Updated the badge for the documentation link in the README file. Now it properly points to the 'echo-artistry'